user authenticated successfully

+++ controllers/expenseController.js
@@ -1,9 +1,11 @@
 const expenseModel = require("../models/expense");
+const User = require('../models/user')
+
 
 exports.expense = async(req,res)=>{
     const {amount,description,category} = req.body
     try{
-       const data = await expenseModel.create({amount,description,category})
+       const data = await expenseModel.create({amount,description,category,userId:req.user.id})
        res.status(201).json({data:data})
     }
     catch(err){
@@ -14,7 +16,7 @@ exports.expense = async(req,res)=>{
 
 exports.getexpense = async(req,res)=>{
     try{
-        const data = await expenseModel.findAll();
+        const data = await expenseModel.findAll({where:{userId:req.user.id}});
         const jsonData = data.map(expense => expense.toJSON());
         res.json(jsonData);        
     }
@@ -24,14 +26,20 @@ exports.getexpense = async(req,res)=>{
 }
 
 exports.delex = async(req,res)=>{
-    const id=req.params.id;
+    const userId=req.user.id;
+    const id = req.params.id
     try{
-        await expenseModel.destroy({where:{id}});
-        console.log('data deleted');
-        res.status(200).json({msg:'data has deleted'})
+      const data = await expenseModel.destroy({where:{userId,id}});
+      if (data > 0) {
+        console.log('Data deleted');
+        res.status(200).json({ msg: 'Data has been deleted' });
+    } else {
+        console.log('No data deleted');
+        res.status(404).json({ msg: 'No matching data found to delete' });
     }
+  }
     catch(err){
         res.status(403).json({msg:"not deleted"})
     }
-}
+  }
